requirements.txt
streamlit
google-generativeai>=0.7.0
pandas
plotly
gspread
oauth2client
streamlit-extras
streamlit-lottie
requests